keep a map of img hashes
If we run this as daemon, we generate lots of (unnecessary) pushes to
the destination.

Keep a simple map of source image hashes, and do actions only when
hashes change.

One change in the semantic is that we need a "data base" of hashes, so
we always pull the image. This is fine, docker already handles pulling
of existing images efficiently.